==> 01_build_simple_dataset.Rmd <==

```{r}
library(tidyverse)
```

load in raw data

```{r}
games <- read_csv("data-raw/games.csv")

atbats <- read_csv("data-raw/atbats.csv")

pitches <- read_csv("data-raw/pitches.csv")

merged_data <- read_csv("data-clean/pitches_with_at_bats.csv")
```

understanding a single game

```{r}
#filtering at bats file for one game

atbats_one_game <- 
	atbats %>% 
	filter(g_id == games$g_id[1]) %>% 
	mutate(g_id = as.character(g_id), ab_id = as.character(ab_id)) %>% 
	select(g_id, ab_id, everything())

#filtering the file merged data for one game

merged_data_one_game <- 
	merged_data %>% 
	filter(g_id == games$g_id[1]) %>% 
	mutate(g_id = as.character(g_id), ab_id = as.character(ab_id)) %>% 
	select(g_id, ab_id, everything())

#filtering the pitches file for one game

pitches_one_game <- 
	pitches %>% 
	filter(ab_id %in% atbats_one_game$ab_id) %>% 
	mutate(ab_id = as.character(ab_id)) %>% 
	select(ab_id, pitch_num, everything())

#filtering the merged data file for 100 games 

merged_data_multiple_games <-
  merged_data %>%
  filter(g_id < 201500100)

#filtering the file merged data for the 2018 season

merged_data_2018 <-
  merged_data %>%
  filter(g_id >= 201800000)

merged_data_100_games <-
  merged_data %>%
  filter(g_id >= 201800000 & g_id <= 201800100)

```

creating the binary dataframe 

```{r}

# creating binary data frame with certain columns from the merged data file
df_final <-
	merged_data %>% 
	select(pitch_type, s_count, b_count, pitch_num, on_1b, on_2b, on_3b, pitcher_id, stand, p_throws) %>% 
	mutate(
		fastball = 
			case_when(
				is.na(pitch_type) ~ "N/A",
				pitch_type == "FF" ~ "yes",
				pitch_type == "FT" ~ "yes",
				pitch_type == "FC" ~ "yes",
				TRUE ~ "no"
			) # Creating a new column which says 'yes' if pitch is fastball, 'no' if not
	) %>% 
    mutate(
    	last_ff_not = 
    		lag(
    			case_when(
    				is.na(pitch_type) ~ "N/A",
    				pitch_type == "FF" ~ "FF",
    				pitch_type == "FT" ~ "FF",
    				pitch_type == "FC" ~ "FF",
    				pitch_type == "CU" ~ "CU/SL",
    				pitch_type == "SL" ~ "CU/SL",
    				pitch_type == "CH" ~ "CH",
    				TRUE ~ "OTHER"), 1
    		) # Adding a column which writes teh value of the previous pitch
    ) %>% 
	select(-pitch_type) # remove the column "pitch_type" as we do not need it anymore

df_final <-
  df_final %>%
    mutate(fastball = ifelse(fastball == "yes", 1, 0))
```


```{r}
df_final_2018 <-
	merged_data_2018 %>% 
	select(pitch_type, s_count, b_count, pitch_num, on_1b, on_2b, on_3b, pitcher_id, stand, p_throws) %>% 
	mutate(
		fastball = 
			case_when(
				is.na(pitch_type) ~ "N/A",
				pitch_type == "FF" ~ "yes",
				pitch_type == "FT" ~ "yes",
				pitch_type == "FC" ~ "yes",
				TRUE ~ "no"
			) # Creating a new column which says 'yes' if pitch is fastball, 'no' if not
	) %>% 
    mutate(
    	last_ff_not = 
    		lag(
    			case_when(
    				is.na(pitch_type) ~ "N/A",
    				pitch_type == "FF" ~ "FF",
    				pitch_type == "FT" ~ "FF",
    				pitch_type == "FC" ~ "FF",
    				pitch_type == "CU" ~ "CU/SL",
    				pitch_type == "SL" ~ "CU/SL",
    				pitch_type == "CH" ~ "CH",
    				TRUE ~ "OTHER"), 1
    		) # Adding a column which writes teh value of the previous pitch
    ) %>% 
	select(-pitch_type) # remove the column "pitch_type" as we do not need it anymore

df_final_2018 <-
  df_final_2018 %>%
    mutate(fastball = ifelse(fastball == "yes", 1, 0))
```

```{r}
df_final_100 <-
	merged_data_100_games %>% 
	select(pitch_type, s_count, b_count, pitch_num, on_1b, on_2b, on_3b, pitcher_id, stand, p_throws) %>% 
	mutate(
		fastball = 
			case_when(
				is.na(pitch_type) ~ "N/A",
				pitch_type == "FF" ~ "yes",
				pitch_type == "FT" ~ "yes",
				pitch_type == "FC" ~ "yes",
				TRUE ~ "no"
			) # Creating a new column which says 'yes' if pitch is fastball, 'no' if not
	) %>% 
    mutate(
    	last_ff_not = 
    		lag(
    			case_when(
    				is.na(pitch_type) ~ "N/A",
    				pitch_type == "FF" ~ "FF",
    				pitch_type == "FT" ~ "FF",
    				pitch_type == "FC" ~ "FF",
    				pitch_type == "CU" ~ "CU/SL",
    				pitch_type == "SL" ~ "CU/SL",
    				pitch_type == "CH" ~ "CH",
    				TRUE ~ "OTHER"), 1
    		) # Adding a column which writes teh value of the previous pitch
    ) %>%
	select(-pitch_type) # remove the column "pitch_type" as we do not need it anymore

df_final_100 <-
  df_final_100 %>%
    mutate(fastball = ifelse(fastball == "yes", 1, 0))

df_final_100 <- df_final_100[-c(1),]
```

Multi Class Dataframe

```{r}

# creating a dataframe for multi classifcation(not binary)
df_multi_class <-
	merged_data_2018 %>% 
	select(pitch_type, s_count, b_count, pitch_num, on_1b, on_2b, on_3b, pitcher_id) %>% 
	mutate(
		fastball = 
			case_when(
				is.na(pitch_type) ~ "N/A",
				pitch_type == "FF" ~ "FF",
				pitch_type == "FT" ~ "FF",
				pitch_type == "FC" ~ "FF",
				pitch_type == "CU" ~ "CU/SL",
				pitch_type == "SL" ~ "CU/SL",
				pitch_type == "CH" ~ "CH",
				TRUE ~ "other" # this column is very similar to the original pitch_type except grouping 4 seam and 2 seam, and curve and slider
			)
	) %>% 
    mutate(
    	last_ff_not = 
    		lag(
    			case_when(
    				is.na(pitch_type) ~ "N/A",
    				pitch_type == "FF" ~ "FF",
    				pitch_type == "FT" ~ "FF",
    				pitch_type == "FC" ~ "FF",
    				pitch_type == "CU" ~ "CU/SL",
    				pitch_type == "SL" ~ "CU/SL",
    				pitch_type == "CH" ~ "CH",
    				TRUE ~ "OTHER"), 1 # adding a column which displays the previous pitch 
    		)
    ) %>% 
	select(-pitch_type) #removing orignial pitch_type column


```


data visualization and analysis

```{r}

pitches_one_game %>% 
  count(pitch_type)
  print(pitch_num)
  
first_pitch <-
  df_final %>% 
    filter(pitch_num == 1)
  
one_strike <-
  df_final%>% 
    filter(s_count == 1)

two_strikes <-
  df_final%>% 
    filter(s_count == 2)

three_balls <-
  df_final%>% 
    filter(b_count == 3)

full_count <-
  df_final%>% 
    filter(b_count == 3 & s_count == 2)


two_strikes %>% 
  count(fastball)

three_balls %>% 
  count(fastball)

full_count %>% 
  count(fastball)

first_pitch %>%
  count(fastball)

df_final %>%
  count(fastball)
```

save dataframe to csv and rds

```{r}
df_final %>% write_rds("data-for-models/df_final.rds")
df_final %>%write_csv("data-for-models/df_final.csv")
df_multi_class %>% write_csv("data-for-models/df_multi_class.csv")
df_final_2018 %>% write_csv("data-for-models/df_final_2018.csv")
df_final_100 %>% write_csv("shiny/df_final_100.csv")

```
