==> 02_data_visualization.Rmd <==

```{r}
library(tidyverse)
```

```{r}

df_final %>%
  count(fastball)
  counts <- table(df_final$fastball)
  barplot(counts, main="Fastballs 2018")
  
ggplot(data = df_final) + 
geom_bar(mapping = aes(x = fastball))

pitches_plot <- within(pitches, 
                   pitch_type <- factor(pitch_type, 
                                      levels=names(sort(table(pitch_type), 
                                                        decreasing=TRUE))))
ggplot(pitches_plot) +
  geom_bar(mapping = aes(x=pitch_type)) 

```
