==> 03_model_RF_python.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd

df_final = pd.read_csv("/Users/shafin/documents/Pitch/data-for-models/df_final.csv") # dataframe of all 3 seasons
df_final_2018 = pd.read_csv("/Users/shafin/documents/Pitch/data-for-models/df_final_2018.csv") # dataframe for 2018 season
df_final_100 = pd.read_csv("/Users/shafin/documents/Pitch/data-for-models/df_final_100.csv") # dataframe for 100 games

# One-hot encode the data using pandas get_dummies
df_final_2018 = pd.get_dummies(df_final_2018)
# Display the first 5 rows of the last 12 columns
df_final_2018.head(5)

from sklearn.model_selection import train_test_split

X=df_final_2018.drop('fastball', axis=1)  # Features
y=df_final_2018.fastball  # Labels

# Split dataset into training set and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3) # 70% training and 30% test

# Import the model we are using
from sklearn.ensemble import RandomForestRegressor
# Instantiate model with 1000 decision trees
rf = RandomForestRegressor(n_estimators = 1000)
# Train the model on training data
rf.fit(X_train, y_train);

y_pred=rf.predict(X_test)

from sklearn import metrics
# Model Accuracy, how often is the classifier correct?
print("Accuracy:",metrics.accuracy_score(y_test, y_pred))
