==> 03_model_svm_python.ipynb <==
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd


df_final = pd.read_csv("/Users/shafin/documents/Pitch/data-for-models/df_final.csv") # dataframe of all 3 seasons
df_final_2018 = pd.read_csv("/Users/shafin/documents/Pitch/data-for-models/df_final_2018.csv") # dataframe for 2018 season
df_final_100 = pd.read_csv("/Users/shafin/documents/Pitch/data-for-models/df_final_100.csv") # dataframe for 100 games

# One-hot encode the data using pandas get_dummies
df_final = pd.get_dummies(df_final)
df_final.head(5)

from sklearn.model_selection import train_test_split

X=df_final.drop('fastball', axis=1)  # Features
y=df_final.fastball  # Labels

# Split dataset into training set and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3) # 70% training and 30% test

#tuning model 
from sklearn.model_selection import GridSearchCV 
from sklearn import svm
from sklearn.svm import SVC 
# defining parameter range 
param_grid = {'C': [0.1, 1, 10, 100, 1000],  
              'gamma': [1, 0.1, 0.01, 0.001, 0.0001], 
              'kernel': ['rbf']}  
  
grid = GridSearchCV(SVC(), param_grid, refit = True, verbose = 3) 
  
# fitting the model for grid search 
grid.fit(X_train, y_train)

from sklearn import svm
clf = svm.SVC(kernel='rbf', C=10, gamma=0.01)
clf.fit(X_train, y_train)
y_pred = clf.predict(X_test)

#Import scikit-learn metrics module for accuracy calculation
from sklearn import metrics
# Model Accuracy: how often is the classifier correct?
print("Accuracy:",metrics.accuracy_score(y_test, y_pred))
